Convert to 64-bit sizes

+++ include/stdio.h
@@ -36,8 +36,8 @@ extern unsigned int getcpuid(void);
 
 /* Formatted input  */
 int _doscan(const char *fmt, va_list ap,
-            int (*getch)(int, int), int (*ungetch)(int, int),
-            int arg1, int arg2);
+            int (*getch)(long, long), int (*ungetch)(long, long),
+            long arg1, long arg2);
 
 int fscanf(int dev, const char *format, ...);
 
@@ -50,7 +50,7 @@ int sscanf(const char *str, const char *format, ...);
 
 /* Formatted output  */
 int _doprnt(const char *format, va_list,
-            int (*putc_func)(int, int), int putc_arg);
+            int (*putc_func)(long, long), long putc_arg);
 
 int fprintf(int dev, const char *format, ...) __printf_format(2, 3);
 int printf(const char *format, ...) __printf_format(1, 2);
@@ -59,7 +59,7 @@ int sprintf(char *str, const char *format, ...) __printf_format(2, 3);
 /* Character and string input and output  */
 int fgetc(int dev);
 char *fgets(char *s, int n, int dev);
-int fputc(int c, int dev);
+int fputc(long c, long dev);
 int fputs(const char *s, int dev);
 
 /** @ingroup libxc */
