+++ system/kprintf.c
@@ -41,7 +41,7 @@ syscall kprintf(const char *format, ...)
 
     va_start(ap, format);
     lock_acquire(serial_lock);
-    retval = _doprnt(format, ap, (int (*)(int, int))kputc, 0);
+    retval = _doprnt(format, ap, (int (*)(long, long))kputc, 0);
     lock_release(serial_lock);
     va_end(ap);
     return retval;
